code/kg-rna-agent/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
biopython>=1.79
requests>=2.26.0
tqdm>=4.62.0
scikit-learn>=1.0.0
flask>=2.0.0
pytest>=6.2.0
jsonschema>=4.0.0
openai>=0.27.0
pyyaml>=6.0
